Day 9 in go
Lots of refactoring as part of this.
Got caught out because we have to deal with mode 2 parameters for write arguments too (was previously presumed to always be mode 0).

I seem to have a bug that the output is not printed before the program exits. Need to work on that! The printOutputs function should stop when the output channel is closed, which happens as soon as execution ends (just before the halt signal is sent). Not sure how to wait before proceeding to the next line.